chore(rig-lancedb): move inline test modules to sibling files
